requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.1
Django==3.0.7
django-crispy-forms==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
